+++ Módulo 03/Interação com API/api-blog/app/Http/Controllers/TagController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Tag;
+use Illuminate\Support\Facades\Validator;
 
 class TagController extends Controller
 {
@@ -14,24 +15,36 @@ class TagController extends Controller
     {
         $tags = Tag::all();
 
-        return response()->json($tags, 201);
+        return response()->json([
+            'mensagem' => 'Tags listadas com sucesso!',
+            'dados' => $tags
+        ], 200);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-    }
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+
+        $validacao = Validator::make($request->all(),[
+            'nome' => 'required|string|max:190'
+        ]);
+
+        if ($validacao->fails()){
+            return response()->json([
+                'mensagem' => 'Erro de validação',
+                'erros' => $validacao->errors()
+            ], 422);
+        }
+
         $tag = Tag::create($request->all());
 
-        return response()->json($tag, 201);
+        return response()->json([
+            'mensagem' => 'Tag cadastrada com sucesso!',
+            'dados' => $tag
+        ], 201);
     }
 
     /**
@@ -40,17 +53,18 @@ class TagController extends Controller
     public function show(string $id)
     {
         $tag = Tag::find($id);
+        if (!$tag){
+            return response()->json([
+                'mensagem' => 'Tag não encontrada!'
+            ], 404);
+        }
 
-        return response()->json($tag, 201);
+        return response()->json([
+            'mensagem' => 'Tag retornada com sucesso',
+            'dados' => $tag
+        ], 200);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-
-    }
 
     /**
      * Update the specified resource in storage.
@@ -59,9 +73,29 @@ class TagController extends Controller
     {
         $tag = Tag::find($id);
 
+        if (!$tag){
+            return response()->json([
+                'mensagem' => 'Tag não encontrada!'
+            ], 404);
+        }
+
+        $validacao = Validator::make($request->all(),[
+            'nome' =>'required|string|max:190'
+        ]);
+
+        if($validacao->fails()){
+            return response()->json([
+                'mensagem' => 'Erro de validação!',
+                'erros' => $validacao-> errors()
+            ], 422);
+        }
+
         $tag->update($request->all());
 
-        return response()->json($tag, 201);
+        return response()->json([
+            'mensagem' => 'Tag atualizada com sucesso!',
+            'dados' => $tag
+        ], 200);
     }
 
     /**
@@ -70,9 +104,14 @@ class TagController extends Controller
     public function destroy(string $id)
     {
         $tag = Tag::find($id);
+        if (!$tag){
+            return response()->json([
+                'mensagem' => 'Tag não encontrada!'
+            ], 404);
+        }
 
         $tag->delete();
 
-        return response()->json('Tag removida com sucesso!', 201);
+        return response()->json(['mensagem' => 'Tag removida com sucesso!'], 201);
     }
 }
